apps/backend/requirements.txt
# Core
flask>=3.0.0
python-dotenv>=1.0.0

# Database
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0

# API & Serialization
flask-marshmallow>=1.2.0
marshmallow-sqlalchemy>=0.30.0
webargs>=8.4.0

# Authentication
flask-jwt-extended>=4.6.0
bcrypt>=4.2.0

# WebSocket (Real-time practice)
flask-socketio>=5.3.0
python-socketio>=5.11.0
eventlet>=0.36.0

# CORS
flask-cors>=4.0.0

# Validation
email-validator>=2.1.0

# Utilities
python-dateutil>=2.8.0

# Audio Analysis
basic-pitch>=0.3.0
librosa>=0.10.0
numpy>=1.24.0
soundfile>=0.12.0

# Development
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.0.0
flake8>=7.0.0

# Production
gunicorn>=21.0.0
